fix egde unsigned long edge cases

+++ Libft/ft_modf.c
@@ -14,7 +14,7 @@ long double ft_modf(long double x, long double *integer)
 {
 	long double int_part;
 
-	int_part = (long double)(int)x;
+	int_part = (long double)(long int)x;
 	*integer = int_part;
 	return (x - int_part);
 }
